refactor(cache): removes global cache
This change removes the global file cache and instead injects the
dependency into every function call that needed access to the cache.
The global cache meant that the test could not be run in parallel i.e.
multiple threads otherwise they would conflict on the global cache.

This change also makes the uri parameter to function a &str instead of a
String. This was because the string was being copied for every call into
the cache for the uri. Now the uri is only copied when it actually
stored in the cache etc.

Finally one test case was remove about multiple packages as it was
asserting internal details of how the router behaves via the global
cache.

Fixes #173