refactor(cache): removes global cache (#179)
* refactor(cache): removes global cache

This change removes the global file cache and instead injects the
dependency into every function call that needed access to the cache.
The global cache meant that the test could not be run in parallel i.e.
multiple threads otherwise they would conflict on the global cache.

This change also makes the uri parameter to function a &str instead of a
String. This was because the string was being copied for every call into
the cache for the uri. Now the uri is only copied when it actually
stored in the cache etc.

Finally one test case was remove about multiple packages as it was
asserting internal details of how the router behaves via the global
cache.

Fixes #173

* fix(lint): update code for various lint fixes

Version 1.47 of Rust found a few places we were using collect and then
immediately iterating over the newly collected vector. The linter
suggested we change to not collect but instead use 'any' to only iterate
once over the initial data.