chore: modify rust toolchain to be more like flux (#212)
In flux, we have a pretty standard process that we use for building rust
code. We deny warnings, we enforce style with fmt checks, and we check
lint with clippy. This patch brings flux-lsp more in-line with flux, and
fixes outstanding issues that occurred as a result of the new stricter
checks.