feat: inherit build setup from flux-build
The newest release of flux, 1.119.0, included changes that won't work on
the old version of rust in the current wasm-build image. That exposed
the problem that we were on an ancient rust. When upgrading, I noticed
that we were still using the `curl | sh` pattern we've been trying to
avoid. The best way to combat that was to just use the `flux-build`
image that we use in flux.

The only remaining need for the wasm-build image was nodejs, which
_also_ used the `curl | sh` pattern, so I took the time to read the
downloaded script and did the tasks (albeit less smart) that the script
did, only manually.