chore: flatten the source tree
In rust, you can have modules as a single file or as a directory, when
the module is too big for a single file. For example, you can have the
`foo` module expressed as `foo.rs` or as `foo/mod.rs`, with other files
in `foo/` being submodules. There isn't a lot of reason to just do
`foo/mod.rs` if the module is _only_ that single file.

This patch simplifies those nested modules to a single file.