fix: update dependencies to most recent versions (#235)
Auditing dependencies, some of these (slightly older) dependencies
allowed for versions of their dependencies with open vulnerabilities. In
general, this didn't necessarily affect this library, because we don't
commit `Cargo.lock` and the default is to take the most up-to-date, but
this change will force updates to stale `Cargo.lock` files on dev
machines.