feat: implement textDocument/formatting
This patch implements `textDocument/formatting` in the new lsp server.
While working on this issue, a problem was discovered in the old
implementation whereby the ranges, likely across the board, were
incorrect. They were left as is, but an explicit assertion on the
existing test was added, with a comment demonstrated its correctness
issues.

Closes #242