chore: reduce the public interface
Building on the knowledge and assertions made by the existence of the
integration tests added in #274, this patch reduces the public interface
provided by the `flux-lsp` lib.

This means that _nothing_ is exposed in the `flux-lsp` lib, but there
are still things exported via `wasm_bindgen`, and they are denoted in
the source properly and enforced with the integration tests.