chore: adjust the integration tests
This code negotiates the differences between the behavior in `master`
and the behaviors expected using the `lsp_types` types. The `lsp_types`
types enforce the spec exactly, but if we've missed out on places to
specify properties, they may be lost in the shuffle to the new
`lspower`-based server.