chore: merge master (#276)
* chore: parallelize ci jobs (#241)

This patch makes the ci jobs parallelized rather than serial. This is a
good idea when there is a credit limit on CircleCI, but the `influxdata`
credit budget allows for parallelizing these tasks.

* chore: Consolidate release scripts and remove unnecessary scripts (#253)

* build(flux): update flux to v0.119.1 (#254)

* chore: Revert USER to UNAME in dockerfile (#255)

* chore: Fix release notes + typos in release script (#256)

* build(flux): update flux to v0.120.1 (#259)

* build(flux): update flux to v0.121.0 (#261)

* chore: Fix release notes in script and remove check for duplicate releases (#263)

* chore: consolidate macros module (#268)

> Flat is better than nested.

This patch merges the 4 lines from the `macros` module into the core
`lib.rs` file and removes the `macros.rs` file.

* build(flux): update flux to v0.123.0 (#272)

* feat: add integration tests (#274)

This patch adds integration tests for the wasm export of the flux lsp
server. See `README.md` for usage.

* chore: reduce the public interface (#275)

Building on the knowledge and assertions made by the existence of the
integration tests added in #274, this patch reduces the public interface
provided by the `flux-lsp` lib.

This means that _nothing_ is exposed in the `flux-lsp` lib, but there
are still things exported via `wasm_bindgen`, and they are denoted in
the source properly and enforced with the integration tests.

* chore: adjust the integration tests

This code negotiates the differences between the behavior in `master`
and the behaviors expected using the `lsp_types` types. The `lsp_types`
types enforce the spec exactly, but if we've missed out on places to
specify properties, they may be lost in the shuffle to the new
`lspower`-based server.

Co-authored-by: Sean Brickley <sean@wabr.io>
Co-authored-by: Adrian Thurston <thurston@colm.net>
Co-authored-by: Christopher M. Wolff <chris.wolff@influxdata.com>
Co-authored-by: Faith Chikwekwe <faithchikwekwe01@gmail.com>