feat: use the builder pattern for LspServer
This patch introduces the builder pattern to the `LspServer`. This work
was a spike in building out a couple of different ways to create an
`LspServer` instance. In the default main case, an `LspServer::new()` is
all that's needed, and doing `LspServer::new(true, None, None, None)` is
less clear about what the random `bool` and `Option` args are.