chore: add some benchmark tests (#303)
This patch adds two benchmark tests, and the underlying setup required
to run those tests. This work was exploratory while chasing some
potential performance issues that I was experiencing in vsflux. It ended
up being informative but not entirely helpful. However, I think it's
still valuable work to commit.

Additionally, a file got reformatted and a visitor was deleted because
it wasn't used.