chore: add a release script
In the same vein as the release script for influxdata/vsflux, this
script does the update of `Cargo.toml` and tag in one step and pushes to
master, causing the CI to then run and publishing to npm.

This script was used to tag/release 0.6.5, so I'm confident it works.