chore: optimize release binary for size
This patch adds release time build options to shrink the size of the
resulting wasm binary. These changes shrink the binary about 2M, so it's
now comparable in size to the wasm binary before the lspower work.

At this point, without going back and re-writing source code, likely in
libflux, we aren't likely to find low-hanging fruit for shrinking the
code size.