feat: log panics from wasm
This patch adds an external dependency that will set up a panic handler
that will log panics originating in wasm to the browser console. This
dependency has worked through support in multiple browsers, so I'm
confident it's going to be better than anything we come up with.

Catching the panic and recovering would be neat, but right now, we get
`RuntimeError: unreachable` which gives us no information and looks like
it's an invocation of `unreachable!`, which has led to some wild goose
chases recently. Having a backtrace would go a long way to fixing
issues.