feat: implement a cleaner wasm interface
This patch implements the interface described in #319, including tests.
It currently puts that work behind a feature flag so that the work isn't
consumable until the bit is flipped for that feature.

Not only does this work provide a cleaner interface, but the resulting
binary is about 20k smaller than the current release binaries.

Closes #319