refactor: Restrict the scope of the store lock
If we keep the lock across an await point we run the risk of deadlocks
so reducing the scope reduce the risk of that. We should perhaps use an
async mutex instead but the mutex is currently also used in sync code so
that is not possible.

Also removes a clone of the entire `store` and instead just clones the
specific file that is used.