feat: no more unwrap/expect/panic (#328)
This patch removes all blind unwrap and panic. This came out of an
iterative need while designing the new API, as I had all sorts issues
where the wasm would stop working because of a panic, and while it was a
coding error, I'd rather get some sort of output.