chore: Add rust toolchain file.
Per <https://rust-lang.github.io/rustup/overrides.html#the-toolchain-file>
`rustup` will check cargo projects for a toolchain file and will fetch
the specified toolchain version if it isn't already present in
`$CARGO_HOME`.

In this, the toolchain file provides a _project-specific_ toolchain
override, allowing individual projects to be pinned to different
versions (if desired).

The intent here is to help unify the standards (lints) applied between
local dev and CI and also signal when a toolchain bump is expected.