feat: buffer the writes and fire on flush
This patch converts the `Outgoing` interface to buffer writes and
then fire the event handlers when `flush` is called. The naive parsing
of bytes to messages has been made a bit more robust, but has a (noted)
glaring issue with it.

Much of this work will likely be heavily refactored/removed after the
fact. The important deliverable here is the external API, and everything
behind-the-scenes can be comfortably re-worked.