fix: don't complete stdlib items starting with underscore
Some functionality was introduced in flux which is...well, sort of
private. It wouldn't be detrimental to know about and use, necessarily,
but we don't want the lsp to help users discover that functionality.
This patch changes the stdlib completion code to not complete items that
start with an underscore.

Fixes #308