test: assert contents not len (#344)
Some of the holdover tests from the previous lsp weren't very specific,
as they asserted on the length of responses rather than contents. This
patch changes that to assert on the contents. "But wait," you might say.
"These tests only assert on labels and names." In these cases, I think
that's okay, as the rest of the data structure isn't really used
(currently) by flux.

Fixes #174