feat: add an LspServerBuilder (#352)
We always used the builder pattern, kinda. Not using a separate struct
certainly carried with it some baggage, which makes it clear why it's
usually implemented that way. This patch rectifies that.

Additionally, it removes some of the options from `LspServerOptions`
because they never should have been there in the first place. Holding on
to secrets and uris isn't really the job of the lsp server, and we
should explicitly forbid that as a pattern.