fix: change javascript API to be send<->receive + onMessage (#342)
Prior to this change the javascript API was a send method with no
response and onMessage handler.
With this change the API is now a send method that returns the
corresponding response and an onMessage handler.

This also removes the need to parse/construct jsonrpc messages with
their Content-Length.

Note no server handler yet sends messages back to the LSP client but the
plumbing is in place now to do so. A follow-on PR will add some initial
diagnostic publishing.