feat: add documentHighlight support (#368)
* feat: add documentHighlight support

With this change the server now supports the
textDocument/documentHighlight feature. The implementation is to report
all references of the identifier at the position.

* fix: refactor common get and parse into single method

Now all instances where we simply needed to read a document from the
store go through the get_document helper method. Additionally all
instances of getting and then parsing a document use a shared helper
method.