feat: remove LspServerBuilder
`LspServerBuilder` was meant to abstract away a bunch of options that we
ported over from the previous lsp implementation. Slowly, those options
have been phased out. The last option left was about disabling folding,
which was odd, because the client should signal whether it supports
folding or not, so there's no need to flip that off and on.

This patch removes the `LspServerBuilder` entirely and just uses the
`LspServer::new` function directly.