feat: implement semantic_tokens_full (#395)
This patch implements `textDocument/semanticTokens/full`. The work here
is mostly in converting ast graph into the lsp's semantic token
data structure. There isn't a 1:1 conversion, and there are a handful of
things that aren't highlighted, but it's a good start. This enables supported
clients to have syntax highlighting with minimal efforts.

Fixes #373