refactor: don't use Result if the Err never matters (#398)
The completion code we have was brought over from the old lsp, and has
some non-idiomatic things in it. It's a bit tangled, but I was trying to
fix some issues in the completion code and find some way to express the
completion code better, and this patch is what came out. It simplifies
things a bit, but isn't at all a final solution. Rather, it's a refactor
of code that should continue to be refactored.