test: fix bench tests
It's clear that the bench tests had not run in quite some time. This
patch fixes them up. Additionally, it adds a CI job to just run the
bench tests, which is more in line with the flux CI that runs the bench
tests just to ensure they actually build.