chore: update to flux 0.153.0 (#399)
This patch updates flux to 0.153.0 - I'm not sure why the release script
didn't do this automatically, but it didn't, so here we are. As an
aside, this new version brought in a change that allowed us to enable
the `lsp` feature in the flux crate, which allowed us to remove a fair
bit of code.