feat: add support for code actions
This patch adds support for the code actions capabilities. The only code
action currently supported is importing a package if an unknown
identifier matches a package that can be imported.

This functionality was then removed from the package code completion
logic, as it wasn't working anyway (it creates broken flux in the web
ui).

Fixes #393