chore: delete the makefile
Makefiles tend to become junk drawers for a project. What's worse is
when they aren't actually tied to the build system, e.g. this project
using `cargo` to handle build. The related code smell is that each make
target ends up being a single command that just calls into a script.
Turns out, that's most of this Makefile, and everything else is
unnecessary.