chore: move server tests
I would almost always prefer a 1000 line file to 10 100 line files,
but I think `server.rs` was big a while ago, and I was mostly just
hoping we'd find a way to shrink it. As we add more and more
capabilities, and with the tests growing at 4x the rate of the server
code itself (a good thing), I'll admit that I was probably dragging me
feet harder than I should have.

This patch moves the tests into its own module.