fix: remove old lsp-related shims (#411)
This patch _started_ by removing the old lsp related shims that existed
everywhere. They no longer are needed, as we get similar logic from
enabling the `lsp` feature of the flux crate.

Like every bit of spaghetti, however, when the strand got pulled, all
sorts of things got tumbled out. The only change that occured was
related to the lsp changes, but there was a lot of moving of code in the
process, to verify what consumed what and from where.