feat: do initial groundwork for multi-file package support (#413)
* feat: add Store for storing files

This patch extracts the mutexes and reference counting out of the
`LspServer` itself. This makes the interface a bit cleaner, and paves
the way for changes to happen under the hood without affecting the
`LspServer` itself.

* feat: introduce a get_package function to Store

This patch adds an API for getting an Flux package from the `Store`
instance. It greatly simplifies much of server code.

Rather than teaching `Store` about `jsonrpc::Error`, a new error enum
was created which can be converted into a `jsonrpc::Error` when they are
exposed to the jsonrpc interface.