chore: disable strict mode by default
This patch removes strict mode by default. In general, I think having a
strict mode is good, and I think that CI should enforce something akin
to strict mode, but I realize now that there is plenty of reason for
that not to get in the way of tinkering during dev. I left it for so
long because go tends to have similar expectations, e.g. unused
variables won't compile, but I never liked that behavior and the man in
the mirror asserting the same things in rust.