feat: shrink the wasm
This patch does a number of things, all in the name of shrinking the
final wasm binary size. The first three are introducing features, two to
change the logging to optional, and one to use a smaller allocator (by
10k).

Additionally, it introduces using some tools to shrink the binary
further by removing more sections that are unneeded. These are used in
`wasm-build.sh`.

Fixes #432