feat: add wasm tools to docker image
This patch adds `wabt` and `binaryen` to the docker image used to build
the wasm binaries. These two packages provide some wasm tools we'll use
in the future to shrink the wasm binary size.

Additionally, the debian version was updated (the stable debian is now
bullseye) and the node version was updated to 16, as 12 goes EOL on
Apr 30, 2022, and 16 is the version used in vsflux.