fix: only show diagnostics relating to the the trigger file
Prior to this patch, there was a bug when, should an error exist
in another file in the package, that diagnostic was reported on _all_
files in the package. This patch fixes that issue to only report errors
on the trigger file, e.g. the file that sent the
`didOpen`/`didChange`/`didSave` event.

Fixes #456