fix: panic when onMessage is called after run (#468)
When calling `run`, the server is consumed and control is, effectively,
lost. When trying to attach an `onMessage` handler after `run`, the
handler can't attach; the `Lsp.processor` is `None` at that point. This
patch causes the wasm to panic explicitly when `onMessage` is called on
a running server.

Fixes #462