fix: cache the lsp::Url in the store (#499)
When the fix for pushing diagnostics out to an entire package went it,
in naively assumed everything was a `file:///` url. The store also
mangled the url past the point of fully reconstructing the url. This
patch caches that url so that it can be referenced again on output of
diagnostics.