fix: unify all the places parsing is implemented
In the early days of the LSP, each rpc implementation would handle its
own parsing (ast or semantic, which is a different issue). Then grew the
necessity of the `Store` implementation, which handled the LSP server
state, at least with regard to files. This seemed a convenient place to
unify the parsing. Unfortunately, not all the parsing was caught in that
first unification (ignoring tests where parsing was used to verify
correctness). This means that if (read: when) we begin caching the ast and
semantic graph, these holdouts will not have the benefit of that
caching.

This patch removes two of the last three holdouts for using the `Store`
implementation for getting parsed flux. The third and final holdout is
in the completion code that is currently being refactored. Rather than
making the change in that branch, it made sense to contain these changes
in their own branch.