fix: fix wasm-bindgen-tests for new formatter
It looks like adding `strict` to the list of features when running
`wasm-pack test` is leaving off the tests in `wasm::tests`. I'm not sure
why that is, but the effect is that there were some failing tests that
were being missed since the most recent flux update.

This patch fixes the tests as well as adjusting the CircleCI tests to
make sure that the tests are being executed properly.