chore: hide lang features behind a feature flag (#525)
* chore: hide lang features behind a feature flag

Months ago, we started thinking about an API for working with flux AST
in javascript. That work then got de-prioritized for other work. This
API is still pretty experimental, and it shouldn't ship in the default
wasm binary for now.

BREAKING CHANGE: The exported `Flux` object in the wasm is no longer
available.

* fix: editor hiccup