refactor: audit and remove unnecessary mutability (#521)
This patch is the result of an audit of all the places where mutable
values are used, and removes them where possible. In some cases, it
wasn't realistic to remove them, either because it would make the code
much less readable, or because that change would be big enough to
warrant its own patch.