refactor: integrate new API and remove lang::get_package_functions
These changes allowed the removal of `lang::get_package_functions` and
related functions.

Aside: changing this, there was a test that never should have worked
that was discovered to be broken. I have no idea what that's about, as
it looks like `csv.from` has never taken a `url` parameter.