fix: address issues in circleci publish config (#534)
There was a typo in the publish config.

While I was in there, I realized that the tests are running on the tag
commit, but they are running in tandem with publish. We want the publish
step to only run if the other bits have completed. As workflows can't
have their own top-level filters, these two workflows were combined,
with a finalizing step of `publish` when the tag matches and all the
other steps have succeeded.