chore: fix changelog generation in release script (#543)
The release script had some issues where it would generate an empty
changelog. I only really noticed when there were a small amount of
changes, but we were trimming off the most recent change thinking it was
the tag update commit (which we don't want). This patch fixes that by
putting that commit _before_ the tag, which is The Right Way™ anyway, as
it means that the tag will contain the correct version number.

Additionally, I added a way to debug the script a little better with the
`DEBUG=1` environment variable, and quieted the ssh support check.