feat: allow initialize to add extra filters
This patch allows initialize to add extra filters atomically on
initialization. This means that users who continue to tack on state
after the composition is turned off will be able to see those changes
when composition is turned back on, as the client has the ability to
specify them all in a single request.