chore: remove collect/sort from `Package.functions`
This patch makes `Package.functions` follow the same API as the rest of
this lang API, in that it returns iterators. In this case, there are
lifetime constraints at the libflux level that prevent returning a raw
iterator (`record` doesn't live long enough), so it still collects the
iterator into a `Vec`, but the function still returns an iterator.

Additionally, it moves the `sort` to the layer where the sort is
required, rather than keeping at this "data access" layer.