fix: use RwLock instead of Mutex for server state
This patch moves the server state from using `Mutex` to using `RwLock`.
The server state is currently unused (it was an idea spike that never
got the rest of its feature fully scheduled). In this case, though,
`RwLock` would be more appropriate, as it would be written once and read
in a number of other places.