chore: remove didSave
In more lsp initialization auditing, we don't signal support for
`textDocument/didSave`. Our `textDocumentSync` capabilities used the
"kind" path rather than the specific options. This patch switches to
using the options, but doesn't use the `textDocument/didSave` event
because it would only do the same as the `textDocument/didChange` event,
which we already handle. This way, we aren't locking the store to set
the same value already set in `textDocument/didChange`.