feat: retain range arguments
This patch enables the user to change the arguments to `range` and have
them be retained by the composition. Prior to this patch, the
composition would just overwrite the `range` arguments to be
`range(start: v.timeRangeStart, stop: v.timeRangeStop)`. Now, if the
user opts to change that range, it is retained over the life of the
composition.