feat: escape schema analysis after other calls (#606)
This patch guards against a situation where a user might add a new
`filter` call that _looks_ like a schema-oriented filter after any
other flux function call.

Flux sees the datastructures as tables with columns. When those tables
come out of storage, we can be confident that any `filter` is operating
on schema-oriented columns. The moment we see a call not is not one of
`from`|`range`|`filter`, we can no longer be confident that the `filter`
call following it is related to schema.